+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## [v3]
 ### Added
 - whitelist-source-range annotation for ingress (2021-01-04)
+- add plugin support (2021-02-05)
+- add Sentry plugin (2021-02-05)
+- add Slack plugin (2021-02-05)